requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
constantly==23.10.4
crc==5.0.0
cryptography==41.0.5
daphne==4.0.0
Django==4.2.7
django-cors-headers==4.3.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Faker==20.0.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.7
Pillow==10.1.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-redis==0.4.0
pytz==2023.3.post1
redis==5.0.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
zope.interface==6.1
